Script run concept.

+++ support/stringa.h
@@ -3,11 +3,11 @@
 #include "array.h"
 
 class stringa : public array {
-	unsigned	find(const char* v, size_t len) const;
-	unsigned	add(const char* v, size_t len);
+	int	find(const char* v, size_t len) const;
+	int add(const char* v, size_t len);
 public:
 	stringa() : array(1) {}
-	unsigned	add(const char* v);
-	unsigned	find(const char* v) const;
-	const char*	get(unsigned v) const { return (v < count) ? (const char*)ptr(v) : ""; }
+	int add(const char* v);
+	int find(const char* v) const;
+	const char*	get(int v) const { return ((unsigned)v < (unsigned)count) ? (const char*)ptr(v) : ""; }
 };
